Replaced Importer with VersionImporter
